boc.pl: make both checkboxes and text input work for config editor
Cunning hack, as 'checked="checked"' string is in TMPL_IF: TMPL_IF
requires a defined value to be true, but then this would appear in text
boxes for which we have no defined value.

By setting the template parameter as '" data-noop="' there is a defined
value for the TMPL_IF check, and the text box value attribute gets
closed with no string and a new attribute with no side-effects is opened
(for the template to close).  Any data-* attribute may be used with
validity under HTML5.