Add ability to do "splits"
New magic account THIS needs supporting elsewhere.

The idea is that a THIS(n) in a TG must always cancel to zero within the
TG.  The 'n' part of 'THIS(n)' is to allow multiple different THISes in
a TG.  'n' shall be specified by the number in Amount (when THIS is the
Creditor) or the debt-share value of THIS otherwise.

Having otherwise meaningful values in these columns does not make sense.

When an 'n' is set in the THIS column, there must either be an Amount
and Creditor *XOR* any number of debtors.

I suppose a transaction between THIS(a) and THIS(b) equates the pools a
and b, but I don't see a need for it, so it'll probably be banned.