TG.pm: make penny allocation more consistent
If two or more accounts equally had the greatest imbalance, the
penalisation would depend on which came out of the hash first.

In the `bod' accounts, this meant that hitting F5 could change the
amounts charged to an account on each go.  Furthermore, since the drain
of kathryn used precomputation, the drain might not match the amount
allocated to the account, leaving an overall balance of +/- a few pence.
Note this did not seem to affect the accounts summary, but even so, it's
a bit crap/unnerving.

sorting the output of the hash makes the penalisation consistent,
against those early in the alphabet -- but the penalty could be positive
or negative, so it's probably fine.

After applying this change all caches need to be nuked, so the precomps
get computed in the same way, otherwise the overall balance will
consistently disagree with the accounts summary!